minor changes made
