+++ src/pages/Hotel/Hotel.jsx
@@ -1,20 +1,7 @@
 import React from "react";
-import "./style.css";
 
-// Components
-import Navbar from "../../components/Navbar";
-import Filter from "../../components/Filter";
-import Footer from "../../components/Footer";
-
-const Hotel = () => {
-  return (
-    <div>
-      <Navbar />
-      <Filter />
-      <div className="hotel-container">Hotel</div>
-      <Footer />
-    </div>
-  );
-};
+function Hotel() {
+  return <div>Hotel</div>;
+}
 
 export default Hotel;
